requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
base58==2.0.1
certifi==2020.12.5
chardet==4.0.0
dateparser==1.0.0
Deprecated==1.2.12
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-filter==2.4.0
django-money==1.3.1
djangorestframework==3.12.4
fusionauth-client==1.27.2
geoip2==4.2.0
gunicorn==20.1.0
idna==2.10
maxminddb==2.0.3
more-itertools==8.7.0
multidict==5.1.0
mysqlclient==2.0.3
pkce==1.0.3
py-bip39-bindings==0.1.6
py-ed25519-bindings==0.1.2
py-moneyed==0.8.0
py-sr25519-bindings==0.1.2
PyMySQL==1.0.2
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
regex==2021.4.4
requests==2.25.1
scalecodec==0.11.2
six==1.16.0
sqlparse==0.4.1
substrate-interface==0.13.2
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
websocket-client==0.58.0
whitenoise==5.2.0
wrapt==1.12.1
xxhash==2.0.2
yarl==1.6.3
